Add Data Processor - Priya

+++ src/main/java/com/thoughtworks/models/StaffSkillSet.java
@@ -5,5 +5,5 @@ import lombok.Getter;
 @Getter
 public class StaffSkillSet {
     String name;
-    Long rating;
+    Double rating;
 }
